final-project-third-iteration-2

+++ src/main/java/dev/nastyazhabko/eventmanager/location/service/LocationService.java
@@ -1,8 +1,10 @@
 package dev.nastyazhabko.eventmanager.location.service;
 
+import dev.nastyazhabko.eventmanager.event.repository.EventRepository;
 import dev.nastyazhabko.eventmanager.location.converter.LocationEntityConverter;
 import dev.nastyazhabko.eventmanager.location.dto.Location;
 import dev.nastyazhabko.eventmanager.location.entity.LocationEntity;
+import dev.nastyazhabko.eventmanager.location.exception.LocationHaveEventsException;
 import dev.nastyazhabko.eventmanager.location.repository.LocationRepository;
 import jakarta.persistence.EntityNotFoundException;
 import org.slf4j.Logger;
@@ -16,10 +18,12 @@ public class LocationService {
     private static final Logger logger = LoggerFactory.getLogger(LocationService.class);
     private final LocationRepository locationRepository;
     private final LocationEntityConverter locationEntityConverter;
+    private final EventRepository eventRepository;
 
-    public LocationService(LocationRepository locationRepository, LocationEntityConverter locationEntityConverter) {
+    public LocationService(LocationRepository locationRepository, LocationEntityConverter locationEntityConverter, EventRepository eventRepository) {
         this.locationRepository = locationRepository;
         this.locationEntityConverter = locationEntityConverter;
+        this.eventRepository = eventRepository;
     }
 
 
@@ -42,6 +46,9 @@ public class LocationService {
         if (!locationRepository.existsById(id)) {
             throw new EntityNotFoundException("Локация с id=" + id + " не найдена");
         }
+        if(eventRepository.existsByLocationId(id)) {
+            throw new LocationHaveEventsException("На локации с id=" + id + " зарегистрированы мероприятия");
+        }
         logger.info("Deleting location with id=" + id);
         locationRepository.deleteById(id);
     }
@@ -58,6 +65,11 @@ public class LocationService {
     public Location updateLocation(int id, Location locationToUpdate) {
         if (!locationRepository.existsById(id)) {
             throw new EntityNotFoundException("Локация с id=" + id + " не найдена");
+        }
+        var oldLocationCapacity = locationRepository.findById(id).get().getCapacity();
+        if (oldLocationCapacity > locationToUpdate.capacity()) {
+            throw new IllegalStateException("Новая вместимость локации должна быть больше чем текущая: "
+                    + oldLocationCapacity);
         }
 
         logger.info("Updating location: id = {}, locationToUpdate = {}", id, locationToUpdate);
